remove requirements.yaml
Collections required for this playbook are pre-installed in collections/

Collections required for z/OS tasks should be in a requirements.yml file next to the playbook, instead of in this repo, because they'll be installed in the wrong place for AAP to see them anyway. It would also be misleading to users, thinking that those collections were already installed.

Signed-off-by: Jacob Emery <jacob.emery@ibm.com>